+++ examples/CustomDensity/src/Data.ts
@@ -39,6 +39,7 @@ namespace LiteMol.Custom {
                 isoSigmaMin: 0,
                 isoSigmaMax: 2,                    
                 radius: 5,
+                showFull: false,
                 minRadius: 0,
                 maxRadius: 10,
                 style: Bootstrap.Visualization.Density.Style.create({