+++ web/Examples/Commands/LiteMol-example.js
@@ -90,6 +90,7 @@ var LiteMol;
                         minRadius: 0,
                         maxRadius: 10,
                         radius: 5,
+                        showFull: false,
                         style: Visualization.Density.Style.create({
                             isoValue: -3,
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
@@ -106,6 +107,7 @@ var LiteMol;
                         minRadius: 0,
                         maxRadius: 10,
                         radius: 5,
+                        showFull: false,
                         style: Visualization.Density.Style.create({
                             isoValue: 3,
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,
@@ -124,6 +126,7 @@ var LiteMol;
                         minRadius: 0,
                         maxRadius: 10,
                         radius: 5,
+                        showFull: false,
                         style: Visualization.Density.Style.create({
                             isoValue: 1.5,
                             isoValueType: Bootstrap.Visualization.Density.IsoValueType.Sigma,